Add game_team_test and game_team. GameTeam is a lightweight team for use by Game.
This will allow changing Game to use a collection of  game_teams instead of assuming only 2 teams team_a and team_b.
Using a collection will eliminate some repeated code and make methods follow DRY.
As a side benefit, it will be easier to generalize a game to involve more than 2 teams. That way the application could be used for a swimming league with swim meets with more than 2 teams.